cloudtests: move cluster related functions to util file
