+++ test/cloudtest/test_disk.py
@@ -10,7 +10,7 @@
 from textwrap import dedent
 
 from materialize.cloudtest.app.materialize_application import MaterializeApplication
-from materialize.cloudtest.k8s import cluster_pod_name
+from materialize.cloudtest.util.cluster import cluster_pod_name
 
 
 def test_disk_replica(mz: MaterializeApplication) -> None:
